# Optimize kernel mode for hybrid QNN test case